+++ scripts/plugins/qatests/testpackages/harperbase/qemuboot.sh
@@ -1,1 +1,0 @@
-../../qemuboot.sh